Update dynamic rules to paramEffects - Update rule makers to effect makers - Reorganise some files - Rename GroupedEffects to EffectGroups and more ---
