+++ package.json
@@ -4,7 +4,7 @@
   "description": "respond to items state in realtime",
   "_main": "src/index.ts",
   "_types": "src/index.ts",
-  "main": "dist/index.js",
+  "main": "src/index.ts",
   "types": "src/index.ts",
   "files": [
     "dist",
