+++ src/declarations.ts
@@ -1,3 +1,5 @@
+import { ParamEffectsGroup } from "./usable/paramEffects";
+
 export interface AllStoreInfoUntyped {
   [StoreName: string]: {
     state: (itemId: any) => any;
@@ -9,7 +11,8 @@ export interface AllStoreInfoUntyped {
 export interface RepondTypesUntyped {
   AllStoreInfo: AllStoreInfoUntyped;
   StepNames: string[] | readonly string[];
-  GroupedEffects: Record<string, Record<string, any>>;
+  EffectGroups: Record<string, Record<string, any>>; // TOTO maybe rename groupedEffects
+  ParamEffectGroups: Record<string, ParamEffectsGroup<any, any>>;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -17,5 +20,3 @@ export interface CustomRepondTypes {}
 
 // The final usable types, with the custom types overriding the default ones
 export interface RepondTypes extends Omit<RepondTypesUntyped, keyof CustomRepondTypes>, CustomRepondTypes {}
-
-// type DDDD = RepondTypes["StepNames"];
