+++ src/index.ts
@@ -19,71 +19,38 @@ export {
   makeEmptyPatch,
   makeMinimalPatch,
   removePartialPatch,
-} from "./usable/compare";
+} from "./usable/patchesAndDiffs";
 
 export { useStore, useStoreEffect, useStoreItem, useStoreItemEffect, useStoreItemPropsEffect } from "./usable/hooks";
 export { addItem, getItem, getPrevState, getRefs, getState, removeItem, setState, onNextTick } from "./usable/getSet";
 
 export {
-  initGroupedEffects,
+  initEffectGroups,
   makeEffect,
   makeItemEffect,
   makeEffects,
   runEffect,
-  runGroupEffects,
-  startAllGroupsEffects as startAllGroupedEffects,
+  runEffectsGroup,
+  startAllEffectsGroups as startAllEffectGroups,
   startEffect,
-  startGroupEffects,
+  startEffectsGroup,
   startNewEffect,
   startNewItemEffect,
-  stopAllGroupsEffects as stopAllGroupedEffects,
+  stopAllEffectsGroups as stopAllEffectGroups,
   stopEffect,
-  stopGroupEffects,
+  stopEffectsGroup,
   stopNewEffect,
 } from "./usable/effects";
 
-// Unique effects helpers
-// TODO update these, but for now keep these so prendy keeps working?
+export { makeParamEffects } from "./usable/paramEffects";
+
 export {
-  makeDynamicRules,
-  makeLeaveRuleMaker,
-  makeNestedLeaveRuleMaker,
-  makeNestedRuleMaker,
-  makeRuleMaker,
-  makeRules,
-} from "./usable/deprecatedRules";
+  makeNestedEffectsMaker,
+  makeEffectsMaker,
+  makeLeaveEffectsMaker,
+  makeNestedLeaveEffectsMaker,
+} from "./usable/effectsMakers";
 
 export type InitialItemsState<T_defaultStateFunctionType extends (...args: any) => any> = {
   [itemId: string]: ReturnType<T_defaultStateFunctionType>; // : AtLeastOne<T>;
 };
-
-// TODO remove these
-export type StoreHelperTypes<
-  T_GetState extends () => { [key: string]: { [key: string]: any } },
-  T_GetRefs extends () => { [key: string]: { [key: string]: any } },
-  T_ItemType extends keyof ReturnType<T_GetState> & keyof ReturnType<T_GetRefs>
-> = {
-  ItemType: keyof ReturnType<T_GetState>;
-  AllItemsState: ReturnType<T_GetState>[T_ItemType];
-  ItemState: ReturnType<T_GetState>[T_ItemType][keyof ReturnType<T_GetState>[T_ItemType]];
-  ItemRefs: ReturnType<T_GetRefs>[T_ItemType][keyof ReturnType<T_GetRefs>[T_ItemType]];
-};
-
-// for generating items with names
-export function makeInitialState({
-  itemPrefix,
-  itemAmount,
-  defaultState,
-}: {
-  itemPrefix: string;
-  itemAmount: number;
-  defaultState: () => any;
-}) {
-  const newInitialState: any = {};
-  // NOTE nowehere's keeping track of the item name number
-  for (var index = 0; index < itemAmount; index++) {
-    const itemId = `${itemPrefix}${index}`;
-    newInitialState[itemId] = defaultState();
-  }
-  return newInitialState;
-}
