+++ src/updating.ts
@@ -2,7 +2,7 @@ import meta, { RecordedChanges, RepondMetaPhase } from "./meta";
 import { forEach } from "chootils/dist/loops";
 import checkEffects from "./checkEffects";
 import { EffectPhase } from "./types";
-import { runNextFrame } from "./settingInternal";
+import { runNextFrame } from "./helpers/frames";
 
 function updateDiffInfo(recordedChanges: RecordedChanges) {
   //  make a diff of the changes
@@ -42,7 +42,7 @@ function runSetStates() {
   meta.setStatesQue.length = 0;
 }
 
-function runAddListeners() {
+function runAddEffects() {
   // adding listeners (rules) are queued and happen here
   // removing listeners happens instantly
 
@@ -54,7 +54,7 @@ function runAddListeners() {
   meta.startEffectsQue.length = 0;
 }
 
-function runListenersWithRunAtStart() {
+function runEffectsWithRunAtStart() {
   for (let index = 0; index < meta.effectsRunAtStartQueue.length; index++) {
     const loopedUpdateFunction = meta.effectsRunAtStartQueue[index];
     loopedUpdateFunction(meta.latestFrameDuration, meta.latestFrameTime);
@@ -153,9 +153,9 @@ function resetRecordedStepChanges() {
 
 function runStepEffects(stepName: string) {
   resetRecordedStepChanges(); // NOTE recently added to prevent derive changes being remembered each time it derives again
-  runListenersWithRunAtStart(); // run the runAtStart listeners
+  runEffectsWithRunAtStart(); // run the runAtStart listeners
   runEffects("duringStep", stepName); //  a running derive-listener can add more to the setStates que (or others)
-  runAddListeners(); // add rules / effects
+  runAddEffects(); // add rules / effects
   runAddAndRemove(); // add and remove items
   runSetStates(); // run the qued setStates
   updateDiffInfo(meta.recordedEffectChanges);
