+++ src/usable/effects.ts
@@ -1,189 +1,15 @@
-import { removeItemFromArrayInPlace } from "chootils/dist/arrays";
-import { breakableForEach, forEach } from "chootils/dist/loops";
-import { default as meta, default as repondMeta } from "../meta";
-import { runWhenDoingEffectsRunAtStart, runWhenStartingEffects, runWhenStoppingEffects } from "../settingInternal";
-import {
-  AllState,
-  DiffInfo,
-  EasyEffect,
-  EasyEffect_Check,
-  Effect,
-  EffectPhase,
-  ItemEffect,
-  ItemType,
-  PropName,
-  RefinedGroupedEffects,
-} from "../types";
-import { getPrevState, getRefs, getState } from "../usable/getSet";
-import { toArray, toMaybeArray } from "../utils";
+import { forEach } from "chootils/dist/loops";
+import { easyEffectToEffect, itemEffectToEffect } from "../helpers/effects/converters";
+import { _startEffect, _stopEffect } from "../helpers/effects/internal";
+import { default as meta } from "../meta";
+import { EasyEffect, Effect, ItemEffect, ItemType, PropName } from "../types";
+import { RepondTypes } from "../declarations";
 
-// -------------------------------------------------------
-// Converters
-// -------------------------------------------------------
-
-// Convert an itemEffect callback to a regular effect callback
-function itemEffectRunToEffectRun<K_Type extends ItemType, K_PropName extends PropName<K_Type>>({
-  check,
-  run,
-}: ItemEffect<K_Type, K_PropName>) {
-  let allowedIdsMap: { [itemId: string]: boolean } | undefined = undefined;
-  const props = toArray(check.prop);
-  const type = check.type;
-  const ids = toMaybeArray(check.id);
-  const becomes = check.becomes;
-
-  if (ids) {
-    allowedIdsMap = {};
-    forEach(ids, (itemId) => {
-      if (allowedIdsMap) allowedIdsMap[itemId] = true;
-    });
-  }
-
-  return (diffInfo: DiffInfo, frameDuration: number, skipChangeCheck?: boolean) => {
-    // if skipChangeCheck is true, it will run the run function regardless of the changes
-    if (skipChangeCheck) {
-      const idsToRun = ids?.length ? ids : meta.itemIdsByItemType[type];
-
-      if (idsToRun?.length) {
-        const prevItemsState = getPrevState()[type] as any;
-        const itemsState = (getState() as AllState)[type];
-        const itemsRefs = getRefs()[type];
-
-        forEach(idsToRun, (itemId) => {
-          breakableForEach(props, (propName) => {
-            const newValue = itemsState[itemId][propName];
-
-            run({
-              itemId: itemId as any,
-              newValue,
-              prevValue: prevItemsState[itemId][propName],
-              itemState: itemsState[itemId],
-              itemRefs: itemsRefs[itemId],
-              frameDuration,
-              ranWithoutChange: true,
-            });
-            return true; // break out of the props loop, so it only runs once per item
-          });
-        });
-      }
-      return true; // return early if skipChangeCheck was true
-    }
-
-    const prevItemsState = getPrevState()[type] as any;
-    const itemsState = (getState() as AllState)[type];
-    const itemsRefs = getRefs()[type];
-    forEach(diffInfo.itemsChanged[type], (itemIdThatChanged) => {
-      if (!(!allowedIdsMap || (allowedIdsMap && allowedIdsMap[itemIdThatChanged as string]))) return;
-
-      breakableForEach(props, (propName) => {
-        if (!(diffInfo.propsChangedBool as any)[type][itemIdThatChanged][propName]) return;
-
-        const newValue = itemsState[itemIdThatChanged][propName];
-
-        let canRunRun = false;
-
-        if (becomes === undefined) canRunRun = true;
-        else if (typeof becomes === "function") {
-          canRunRun = becomes(newValue, prevItemsState[itemIdThatChanged][propName]);
-        } else if (becomes === newValue) canRunRun = true;
-
-        if (!canRunRun) return;
-
-        run({
-          itemId: itemIdThatChanged as any,
-          newValue,
-          prevValue: prevItemsState[itemIdThatChanged][propName],
-          itemState: itemsState[itemIdThatChanged],
-          itemRefs: itemsRefs[itemIdThatChanged],
-          frameDuration,
-          ranWithoutChange: false,
-        });
-        return true; // break out of the loop, so it only runs once
-      });
-    });
-  };
-}
-
-function easyEffectCheckToEffectChecks<K_Type extends ItemType>(effectCheck: EasyEffect_Check<K_Type>) {
-  const checksArray = toArray(effectCheck);
-
-  return checksArray.map((check) => ({
-    types: toMaybeArray(check.type),
-    names: toMaybeArray(check.id),
-    props: toMaybeArray(check.prop),
-    addedOrRemoved: check.addedOrRemoved,
-  }));
-}
-
-export function easyEffectToEffect<T_EasyEffect extends EasyEffect<any>>(easyEffect: T_EasyEffect): Effect {
-  return {
-    ...easyEffect,
-    id: easyEffect.id ?? toSafeEffectId("effect"),
-    checks: easyEffectCheckToEffectChecks(easyEffect.check),
-  };
-}
-
-export function itemEffectToEffect<K_Type extends ItemType, K_PropName extends PropName<K_Type>>(
-  itemEffect: ItemEffect<K_Type, K_PropName>
-): Effect {
-  let effectName = itemEffect.id || "unnamedItemEffect" + Math.random();
-
-  return easyEffectToEffect({
-    ...itemEffect,
-    id: effectName,
-    check: { ...itemEffect.check, becomes: undefined, prop: toMaybeArray(itemEffect.check.prop) },
-    run: itemEffectRunToEffectRun(itemEffect),
-  });
-}
-
-// --------------------------------------------------------------------
-// Internal functions
-// --------------------------------------------------------------------
-
-export function _startEffect(newEffect: Effect) {
-  const phase: EffectPhase = !!newEffect.atStepEnd ? "endOfStep" : "duringStep";
-  const step = newEffect.step ?? "default";
-
-  if (newEffect.runAtStart) {
-    runWhenDoingEffectsRunAtStart(() => newEffect.run(meta.diffInfo as any, 16.66666, true /* ranWithoutChange */));
-  }
-
-  runWhenStartingEffects(() => {
-    meta.allEffects[newEffect.id] = newEffect as unknown as Effect;
-    if (!meta.effectIdsByPhaseByStep[phase][step]) {
-      // add the effect to a new array
-      meta.effectIdsByPhaseByStep[phase][step] = [newEffect.id];
-    } else {
-      if (!meta.effectIdsByPhaseByStep[phase][step]?.includes(newEffect.id)) {
-        // add the effect to the array if it's not already there
-        meta.effectIdsByPhaseByStep[phase][step].push(newEffect.id);
-      }
-    }
-  });
-}
-
-export function _stopEffect(effectName: string) {
-  runWhenStoppingEffects(() => {
-    const effect = meta.allEffects[effectName];
-    if (!effect) return;
-    const phase: EffectPhase = !!effect.atStepEnd ? "endOfStep" : "duringStep";
-    const step = effect.step ?? "default";
-
-    removeItemFromArrayInPlace(meta.effectIdsByPhaseByStep[phase][step] ?? [], effect.id);
-
-    delete meta.allEffects[effectName];
-  });
-}
-
-export function toSafeEffectId(prefix?: string): string {
-  const counterNumber = repondMeta.autoEffectIdCounter;
-  repondMeta.autoEffectIdCounter += 1;
-  return (prefix || "autoEffect") + "_" + counterNumber;
-}
-
-// --------------------------------------------------------------------
-// Usable functions
-// --------------------------------------------------------------------
+// Helper type to strip "Effects" suffix from group names
+type RemoveEffectsSuffix<T extends string> = T extends `${infer Prefix}Effects` ? Prefix : T;
+export type RefinedEffectGroups = {
+  [K in keyof RepondTypes["EffectGroups"] as RemoveEffectsSuffix<K>]: RepondTypes["EffectGroups"][K];
+};
 
 export function startNewEffect<K_Type extends ItemType>(theEffect: EasyEffect<K_Type>) {
   _startEffect(easyEffectToEffect(theEffect));
@@ -204,52 +30,52 @@ export function stopNewEffect(effectName: string) {
 
 // This is really startGroupedEffect
 export function startEffect<
-  K_EffectGroup extends keyof RefinedGroupedEffects,
-  K_EffectName extends keyof RefinedGroupedEffects[K_EffectGroup] & string
+  K_EffectGroup extends keyof RefinedEffectGroups,
+  K_EffectName extends keyof RefinedEffectGroups[K_EffectGroup] & string
 >(groupName: K_EffectGroup, effectName: K_EffectName) {
-  const theEffect = (meta.allGroupedEffects as any)[groupName][effectName];
+  const theEffect = (meta.allEffectGroups as any)[groupName][effectName];
   if (!theEffect) return console.warn("no effect found for ", groupName, effectName);
   _startEffect(theEffect);
 }
 
 export function stopEffect<
-  K_EffectGroup extends keyof RefinedGroupedEffects,
-  K_EffectName extends keyof RefinedGroupedEffects[K_EffectGroup] & string
+  K_EffectGroup extends keyof RefinedEffectGroups,
+  K_EffectName extends keyof RefinedEffectGroups[K_EffectGroup] & string
 >(groupName: K_EffectGroup, effectName: K_EffectName) {
-  const theEffect = (meta.allGroupedEffects as any)[groupName][effectName];
+  const theEffect = (meta.allEffectGroups as any)[groupName][effectName];
   if (!theEffect) return console.warn("no effect found for ", groupName, effectName);
   _stopEffect(theEffect.id);
 }
 
-export function startGroupEffects<K_EffectGroup extends keyof RefinedGroupedEffects>(groupName: K_EffectGroup) {
-  const theGroup = (meta.allGroupedEffects as any)[groupName];
+export function startEffectsGroup<K_EffectGroup extends keyof RefinedEffectGroups>(groupName: K_EffectGroup) {
+  const theGroup = (meta.allEffectGroups as any)[groupName];
   forEach(Object.keys(theGroup), (effectName) => startEffect(groupName, effectName));
 }
 
-export function stopGroupEffects<K_EffectGroup extends keyof RefinedGroupedEffects>(groupName: K_EffectGroup) {
-  const theGroup = (meta.allGroupedEffects as any)[groupName];
+export function stopEffectsGroup<K_EffectGroup extends keyof RefinedEffectGroups>(groupName: K_EffectGroup) {
+  const theGroup = (meta.allEffectGroups as any)[groupName];
   forEach(Object.keys(theGroup), (effectName) => stopEffect(groupName, effectName));
 }
 
-export function startAllGroupsEffects() {
-  forEach(Object.keys(meta.allGroupedEffects), (groupName) => startGroupEffects(groupName));
+export function startAllEffectsGroups() {
+  forEach(Object.keys(meta.allEffectGroups), (groupName) => startEffectsGroup(groupName));
 }
 
-export function stopAllGroupsEffects() {
-  forEach(Object.keys(meta.allGroupedEffects), (groupName) => stopGroupEffects(groupName));
+export function stopAllEffectsGroups() {
+  forEach(Object.keys(meta.allEffectGroups), (groupName) => stopEffectsGroup(groupName));
 }
 
 export function runEffect<
-  K_EffectGroup extends keyof RefinedGroupedEffects,
-  K_EffectName extends keyof RefinedGroupedEffects[K_EffectGroup] & string
+  K_EffectGroup extends keyof RefinedEffectGroups,
+  K_EffectName extends keyof RefinedEffectGroups[K_EffectGroup] & string
 >(groupName: K_EffectGroup, effectName: K_EffectName) {
-  const theEffect = (meta.allGroupedEffects as any)[groupName][effectName];
+  const theEffect = (meta.allEffectGroups as any)[groupName][effectName];
   if (!theEffect) return console.warn("no effect found for ", groupName, effectName);
   theEffect.run(meta.diffInfo as any, 16.66666, true /* ranWithoutChange */);
 }
 
-export function runGroupEffects<K_EffectGroup extends keyof RefinedGroupedEffects>(groupName: K_EffectGroup) {
-  const theGroup = (meta.allGroupedEffects as any)[groupName];
+export function runEffectsGroup<K_EffectGroup extends keyof RefinedEffectGroups>(groupName: K_EffectGroup) {
+  const theGroup = (meta.allEffectGroups as any)[groupName];
   forEach(Object.keys(theGroup), (effectName) => runEffect(groupName, effectName));
 }
 
@@ -274,7 +100,7 @@ export function makeEffects<K_EffectName extends string>(
   return effectsToAdd({ itemEffect: makeItemEffect, effect: makeEffect });
 }
 
-export function initGroupedEffects<T extends Record<string, ReturnType<typeof makeEffects>>>(groups: T): T {
+export function initEffectGroups<T extends Record<string, ReturnType<typeof makeEffects>>>(groups: T): T {
   const transformedGroups: Record<string, ReturnType<typeof makeEffects>> = {};
 
   Object.entries(groups).forEach(([key, value]) => {
@@ -296,7 +122,7 @@ export function initGroupedEffects<T extends Record<string, ReturnType<typeof ma
   });
 
   // Store the transformed groups
-  meta.allGroupedEffects = transformedGroups as any;
+  meta.allEffectGroups = transformedGroups as any;
 
   return groups;
 }
