+++ src/usable/patchesAndDiffs.ts
@@ -1,6 +1,6 @@
 import { getUniqueArrayItems } from "chootils/dist/arrays";
 import { forEach } from "chootils/dist/loops";
-import { addItem, getDefaultStates, getItemTypes, removeItem, setState } from "../usable/getSet";
+import { addItem, getDefaultStates, getItemTypes, removeItem, setState } from "./getSet";
 import { createDiffInfo } from "../getStatesDiff";
 import meta, { UntypedDiffInfo, initialRecordedChanges } from "../meta";
 import { AllState, DiffInfo, GetPartialState, ItemIdsByType, ItemId, ItemType, PropName } from "../types";
